my_utils/pyproject.toml
[tool.poetry]
name = "my-utils" # 패키지 이름은 임포트할 때 사용할 이름과 일치시키는 것이 좋습니다.
version = "0.1.0" # 필요에 따라 버전 수정
description = "Utility functions and modules for various projects." # 프로젝트 설명
authors = ["Your Name <you@example.com>"] # 실제 정보로 수정해주세요.
readme = "README.md" # 필요시 README.md 파일을 해당 디렉토리에 생성하세요.

# 실제 파이썬 코드가 있는 디렉토리를 지정합니다.
# 이 경우 src/my_utils 안에 코드가 있으므로 다음과 같이 설정합니다.
packages = [{include = "my_utils", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10" # 마루님께서 사용하시는 파이썬 버전과 일치시키거나 호환되도록 설정

# my_utils 자체에서 사용하는 다른 라이브러리가 있다면 여기에 추가합니다.
# 예:
# PyYAML = ">=6.0"
# numpy = "^1.26.0"
pyyaml = "^6.0.2"
dlib = "^20.0.0"
numpy = "^1.26.0"
opencv-python = "^4.11.0.86"
torch = "^2.7.1"
ultralytics = "^8.3.158"
pillow = "^11.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

